14 febrero

+++ cub.h
@@ -6,7 +6,7 @@
 /*   By: jgoikoet <jgoikoet@student.42urduliz.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/30 16:06:51 by jgoikoet          #+#    #+#             */
-/*   Updated: 2024/02/13 17:45:38 by jgoikoet         ###   ########.fr       */
+/*   Updated: 2024/02/14 12:27:55 by jgoikoet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -88,7 +88,7 @@ typedef struct s_data
 	t_img	active;
 	void	*norte;
 
-	int 	correct;
+	double 	correct;
 }				t_data;
